+++ systemimage/config.py
@@ -163,4 +163,5 @@ class _Proxy:
     def __delattr__(self, name):
         delattr(_config, name)
 
+
 config = _Proxy()